pyproject.toml
[tool.poetry]
name = "ccx"
version = "0.2.0"
description = "Claude Code eXtension — project-aware development pipeline"
authors = ["Your Name <you@example.com>"]
packages = [{include = "ccx", from = "src"}]
include = ["ccx/skills/**/*.md", "ccx/base-context.example.yaml", "ccx/hooks/**/*.sh"]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
click = "^8.0"
mcp = {version = ">=1.0", extras = ["cli"]}
pathspec = "^0.12"

[tool.poetry.scripts]
ccx = "ccx.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
